Remove CI/CD build step and reorder lint and tests
The app will already try to build during the tests and the promotion step if it is promoted. This should reduce the CI/CD time by 2 minutes. Also moved swiftlint to the first step which should cause the CI/CD to fail faster if there are lint issues. This should also reduce build time when lint issues exist.