Fix camera never appearing on allowing camera permissions
Fix the app never fully starting on first load after clicking allow camera. There is still an issue of it taking ~10 seconds for the camera to appear but it seems this is due to the first Vision call causing some background metal process to fail twice before resetting a cache.