Slightly increase tappable area for playback submit
Slightly adds spacing to the playback view so that the submit button has room to be tapped. It was being overlayed by the LoopingPlayers and I could not figure out how to get a zIndex to properly put the button on top. Use Debug -> View -> Capture View Hier.  to see.